+++ src/background/seelen.rs
@@ -15,6 +15,7 @@ use crate::{
     error_handler::Result,
     hook::register_win_hook,
     log_error,
+    modules::uwp::UWP_MANAGER,
     monitor::Monitor,
     seelen_shell::SeelenShell,
     seelen_weg::SeelenWeg,
@@ -91,7 +92,7 @@ impl Seelen {
         }
 
         self.ensure_folders()?;
-        self.load_uwp_apps_info()?;
+        UWP_MANAGER.lock().refresh()?;
 
         let data_path = app.path().app_data_dir()?;
         self.state = State::new(&data_path.join("settings.json")).unwrap_or_default();
@@ -254,61 +255,6 @@ impl Seelen {
                 Err(err) => log::error!("schedule auto start Failed to wait for process: {}", err),
             };
         });
-
-        Ok(())
-    }
-
-    fn load_uwp_apps_info(&self) -> Result<()> {
-        let time = std::time::Instant::now();
-        let pwsh_script = include_str!("load_uwp_apps.ps1");
-        let pwsh_script_path = temp_dir().join("load_uwp_apps.ps1");
-        std::fs::write(&pwsh_script_path, pwsh_script).expect("Failed to write temp script file");
-
-        let manifest_path = self
-            .handle()
-            .path()
-            .app_data_dir()?
-            .join("uwp_manifests.json");
-        let manifest_path_exist = manifest_path.exists();
-
-        let task = async move {
-            let result = get_app_handle()
-                .shell()
-                .command("powershell")
-                .args([
-                    "-ExecutionPolicy",
-                    "Bypass",
-                    "-NoProfile",
-                    "-File",
-                    &pwsh_script_path.to_string_lossy(),
-                    "-SavePath",
-                    manifest_path
-                        .to_string_lossy()
-                        .trim_start_matches("\\\\?\\"),
-                ])
-                .status()
-                .await;
-
-            let duration = time.elapsed();
-            match result {
-                Ok(status) => log::trace!(
-                    "load_uwp_apps took: {}s, exit code: {}",
-                    duration.as_secs_f32(),
-                    status.code().unwrap_or_default()
-                ),
-                Err(err) => log::error!(
-                    "load_uwp_apps took: {}, failed to wait for process: {}",
-                    duration.as_secs_f32(),
-                    err
-                ),
-            };
-        };
-
-        if !manifest_path_exist {
-            tauri::async_runtime::block_on(task);
-        } else {
-            tauri::async_runtime::spawn(task);
-        }
 
         Ok(())
     }